test_wipe_lnx.sh: use --null-data for grep to reduce memory consumption
Might be necessary to test with larger image sizes while having low
free memory (e.g. to test BTRFS with an image size of 1GB in order to
test without the feature mixed-bg).

In addition to reducing memory consumption it makes the output of grep
nicier if not redirected to /dev/null.